+++ routes/index.routes.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const membersRouter = require("./members");
+const membersRouter = require("./members.routes");
 const reservationRouter = require("./reservation.routes");
 
 router.use("/members", membersRouter);
