+++ routes/members.routes.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 
-const MembersController = require("../controller/members");
+const MembersController = require("../controller/members.controller");
 const membersController = new MembersController();
 
 router.post("/signup", membersController.createMembers);
